Initialize only once per sanctions/iban service instance (#3602)
* Initialize only once per sanctions/iban service instance

Previously we early return if apiKey is not empty. But if have many
concurrent caller of the service, we can still end up fetching the api key
multiple times.
Instead of checking apiKey only, we ensure the initialization is actually
invoked only once per instance. After the first caller calls `initialize`,
we set the instance's `initializationPromise` and the next caller just awaits
`initializationPromise`.