Direction-less entity variable  (#3593)
* Support for direction-less entity variable

Previously, users could only choose one of sender or receiver. For example,
we needed to choose between "origin transaction amount" and "destination transaction amount".
So, if we wanted to have logic to check both directions, we would need to construct the logic to
be - "origin transaction amount > 100" OR "destination transaction amount > 100".
If we didn't care about the direction, it was cumbersome to configure the logic as there would be
many OR conditions.

Now, for each entity variable with direction, we create one more variable "without" direction.
(e.g "TRANSACTION:userId__BOTH", "BUSINESS_USER:ageYears__BOTH") and when such variable is
used in logic, we transform the logic to replace the direction-less variables with directional
variables and add OR conditions. For example, if the logic is - "TRANSACTION:userId__BOTH" == '1',
the transformed logic will be - "TRANSACTION:originUserId == '1' OR "TRANSACTION:destinationUserId" == '1'.
This will simplify the logic a lot if users want to check both directions. If users only want to check
a single direction, they can still use the directional variables as before.

* Don't show direction-less variables in the "aggregate Field"

* Handle both lhs and rhs are direction-less variables