Enable UNIQUE_VALUES aggregation variable to be used in RHS (#3623)
* Enable UNIQUE_VALUES aggregation variable to be used in RHS

To support the logic "origin iban IN unique ibans in the past", the
UNIQUE_VALUES aggregation variable needs to be in `text` type (because
lhs field and rhs field need to have the type).
When a UNIQUE_VALUES agg var is created, we'll create one var with `multiselect` type
and one var with `text` type. The definitions of these two vars are always the same,
and users will only see the first one in the agg vars list. For the first var (with
key ending with $1), it can only be used in LHS. For the second var (with key ending
with $2), it can ony be used in RHS if LHS field is `text` type.

Example rule logic:
```
{
    "and":
    [
        {
            "in":
            [
                {
                    "var": "TRANSACTION:originPaymentDetails-IBAN"
                },
                {
                    "var": "agg:69634ef8-7fb0-4247-89af-580a6ba4fa0d$2"
                }
            ]
        },
        {
            "some":
            [
                {
                    "var": "agg:69634ef8-7fb0-4247-89af-580a6ba4fa0d$1"
                },
                {
                    "in":
                    [
                        {
                            "var": ""
                        },
                        [
                            "iban1",
                            "iban2"
                        ]
                    ]
                }
            ]
        }
    ]
}
```

* add validation