V8: support aggregation group by field (#3665)
* Update default values of userDirection and transactionDirection

* V8: support new aggregate group by field

Add the new optional field `aggregationGroupByFieldKey` field to `RuleAggregationVariable`.
Without it, we save the aggregation data like `{ value: 100 }`. With it, we save it like
`{ value: { group1: 10, group2: 90 } }`.
When verifying a transaction, we get the group value according to the group field of the transaction.
For example, if we group by transaction type and the aggregation data is
`{ value: { TRANSFER: 10, DEPOSIT: 20 } }` and the incoming transaction is of type TRANSFER, we'll
use the value `10` as the value of the aggregation variable.

* Add logs to USER_ONGOING_SCREENING type rule

* Add aggregation variable summary

To make it easier to understand an aggregation variable

* V8: disable second/minute granularity and only allow integer value

* fix auto aggregation variable name