Always use the latest aggregation data in sandbox (#3796)
For performance concerns we don't rebuild the aggregation data when verifying
a transaction. So when a rule is changed, new transaction might not be correctly
hit because the latest aggregation data is not ready. This could cause confusion
when customers are evaluating rules in sandbox manually.
Change to always use the latest aggregation data in sandbox to avoid false negatives.