Fix failing to update 'AI Sources' (#3888)
We were passing the whole tenant settings object when updating 'AI Sources', causing
the permission error as some setting properties can only be set by root user.

Change to only pass 'aiSources' from Console. And in the backend we compute the diff
object first, so it won't fail if we pass the whole object again.