Fix dashboard-hits-by-user-stats-hourly collection missing data (#3904)
* Fix dashboard-hits-by-user-stats-hourly collection missing data

For a given time range, we refresh case stats and transaction stats concurrently,
but for `hits-by-user-stats` stats, the cases stats and the transaction stats stay
in the same collection with the same key (date, direction, userId). If case stats
is refreshed later than transaction stats, it could incorrectly clean up the transaction
stats if there're no cases in the target time range.

To fix this issue, for case stats, we only clean up the stale stats if `casesCount` is
present. Likewise, for transaction stats, we only clean up the stale stats if `rulesRunCount`
is present, then we won't wrongly remove records.