V8: Enable comparing 'text' variable to 'select' variable (#3914)
* V8: support comparing 'select' variable with 'text' variable

In 'react-awesome-query-builder', only the variables with the same type can be
compared, so 'select' variable cannot compare with 'text' variable.
As they're both just "text" value type under the hood, and we need to support
comparing 'select' variable with unique values aggregation variable, we change
all 'select' variables to 'text' type.
We still use select/multiselect custom widget to render a select/multiselect variable
although it's marked as text type, and we still show select-only operators if LHS is
a select variable. The only behavioral change is that we can now compare a text variable
with a select variable (vice versa).

* Update logic when config is changed

We update the logic when variables are updated. Because when a variable is
updated, it might not be valid anymore in the logic. If we don't updat the logic
from state.tree, the json logic could contain invalid conditions which will lead
to failing to save a rule due to the backend validation.