Update notion ticket status to 'Ready to Test' after sandbox deployment (#3928)
After a change is deployed to sandbox, we change the corresponding ticket
status to 'Ready to Test'.
After a change is deployed to prod, we change the ticket status to 'Done (Archive').
Previously, we test again a ticket when it's in 'Done (Weekly)` and manually move to 'Done (Archive)'. Now we'll do test testing when a ticket is in 'Ready to Test' status (deployed to sandbox).