Show new case updates message when a case is updated by other people (#3925)
In 19c7f8ca, we refetch a case every 60 seconds and display the latest case data when there're
new update. This feature was disabled in 026b070 as it'll close a open modal on re-rendering.
Instead of showing the latest data automatically, we now just show an alert message when there're
updates, and a user needs to refresh the page to see the latest data.
A benefit of this approach is that users are aware of new updates and have control to fetch the
latest data or not.
We could support proper real-time updates later in the future.