Fix confirmation modal closed after rule logic config prefetch complete (#3985)
We use `useRuleLogicConfig()` to prefetch rule logic config in the rule
library/my rules page. But after the prefetch is done, the page will be
re-rendered because `onChange`, `items` are new objects. If a modal is open,
it'll be closed during the re-render, causing e2e tests to fail.
Fix by making `onChange`, `items` memoized.