Transaction event variables and Transaction entity variable filters (#4223)
* Save rules result in a user event
To be the same as transaction event.

* Add filtersLogic to RuleEntityVariableInUse

* remove old unused code

* Add tx event variables / support tx entity variable filters
For a transaction/transaction event entity variable, if filters not defined,
we use the latest transaction/transaction event. If filters are defined, we
find the matched transaction event from the lastest one to the earliest one,
and early return on match.

* Add 'Timestamp diff (seconds)' function

* Function groups - precursor
renderFunc needs to support option groups first

* Refactor and support duplicating entity variables

Make `key` and `entityKey` required for `RuleEntityVariableInUse`.

* Change entity/aggregation variable key to use short ID

* Handle transaction event variable in an aggregation variable filters logic
If transaction event entity variables are being used in the filters logic of
an aggregation variable, we use a transaction's latest transaction event.
We already load transaction events of a transaction before running a rule, so
no eatra query is needed when running a rule. However, when we rebuild an
aggregation variable, we need to load the latest transaction event for each
historical transaction as the retrieved transactions don't include event data.
This can be optimized later to have more efficient rebuilding.