Fix sanctions search usages metric (#4256)
After 7622b16, we started to have duplicated search records even if we only call
ComplyAdvantage once (will be fixed by FR-5132).
Change the way we count the number of searches to count the unique number of
CA search IDs instead. Then we'll have the correct metric even if we have the same
regression again.