Fix v8 pre-aggregation issues (#4310)
* Increase Fargate job retry attempts from 2 to 15

* Create CloudWatch alarm for failed step functions

* Insert a new job with PENDING status in jobTriggerHandler instead

* Improve v8 aggregation logs

* Fix job.metadata.tasksCount more than job.metadata.completeTasksCount

It could happen when unique values aggregation variable is used as we have two identical
variables in the list ($1, $2). Change to ensure tasksCount is always the same as the
unique count of message deduplication IDs.

* Update the aggregated data for transactions created during the deployment of the rule

Beacuse we only change a rule instance from Deploying to Active if all the pre-aggregation tasks
are done, say we've pre-aggregated for user 1 at t1 and we'll pre-aggregate for all users
at t10, new transactions from user 1 created between t1 and t10 won't be considered in the
aggregation data.
Fix by updating aggregation data for deploying rule instances for new transactions.
We do this asynchronously in the tx consumer. And change to save transaction event before
saving transaction to make unit test simpler (when running unit tests, we run the tx consumer
synchronously when saving a transaction).

* Alert if rule instance is stuck in deploying state for more than 24 hours