Fix unexpectedly created cases triggered by user rules (#4363)
* Don't update user executedRules/hitRules for ongoing screening user rules

In 88f152f, we create cases when inside the ongoing screening user rule job instead
of updating user.executedRules/hitRules. It should've been applied to 7425202 as well.

* Avoid creating unexpected new cases for users

This is a workaround to avoid creating redundant cases. In 748200a, we update
user.riskLevel in DynamoDB, but if a case was created for rule A and was closed afterwards,
updating user.riskLevel alone will trigger a new case which is unexpected.
We only want to create a new case if the user details have changes
(when there're changes, we'll run user rules again).

* Avoid creating new alert/case if sanction hit IDs is empty

After the previous fixes, we probably don't need to do this. But this
could be a safety net if things go wrong again, then we don't create cases
we shouldn't create.