Optimize saving aggregation values if group-by is used (#4406)
* Reduce aggregation rebuild read timeout to 10 min

As write might need more time.

* Set aggregation ready marker ttl to 1 year

If an aggregation variable is live for over 1 year, we'll rebuild once per
year. If an aggregation variable is changed, the old ready marker will be
removed after 1 year.

* Optimize saving aggregation values if group-by is used

Follow-up of 0b0dec0.
- Deduplicate groups
- Gather all the write requests first then do bulk writes to have
  the minimal number of dynamodb put requests.

* Don't save dynamodb partition keys if ttl exists

If 'ttl' exists, it'll be auto-deleted by DynamodB, so we don't
need to track it.