Reverify transactions (improved) (#4438)
* Reverify transactions using batch job runner (fargate)

Define `afterTimestamp`, `beforeTimestamp`, `extraFilters` to filter the
transactions to reverify.
Define `ruleInstanceIds` to only reverify transactions with the specified rules.

```
type ReverifyTransactionsBatchJobParameters = {
  afterTimestamp: number
  beforeTimestamp: number
  ruleInstanceIds: string[]
  extraFilter?: Filter<InternalTransaction>
}
```

Follow-up of e2d0e3b3

* Don't create a new alert if a (tx,rule) exists in an existing alert

For example, we have 2 rules - R-1, R-2. When a transaction is created, R-1 is
hit and alert A-1 is created. Then the case is closed. Later a transaction event
for the same transaction is sent and hit R-2. Previously we'll create 2 alerts -
one for R-1, one for R-2. But we shouldn't create the one for R-1 again as it was
already closed before. Now we'll only create a new alert for R-2.