Double check new rule instance ID before saving it (#4760)
We noticed that 3 prod rule instances (all created between 2024-04-17 to 2024-04-20)
from different tenants have the incorrect counter which will lead to existing rule to be
overwritten if duplicating the rule instance. I didn't bother to find the root cause.

Add an extra safety net by checking the newly generated rule instance ID to see if there's
an existing rule instance with the same ID. If true, we increment the counter one more time,
if same thing happens, we throw an error as something is seriously wrong.