Limit the max active Console sessions per user (#4882)
* Limit the max active Console sessions per user

Add a new tenant setting - maxActiveSessions. When it's defined (greater than 0),
for each request, we check if there's an active session present. If not, we throw
Unauthorized and the user will need to log in again (if maxActiveSessions not
configured, it's the same as the old behaviour).
After login, we redirect to '/post-login' and we add a new entry to the user's active
sessions (and kick out the old ones if there're more than the limit, fifo). Once done,
we redirect user to the home page.

* include device fingerprint