PNB backfill phase 2 preparation (#5021)
* Allow verifying an existing transaction if validateTransactionId is false

We already backfilled all the transactions from PNB without running rules, and
now we want to verify the last 2-week transactions.

* Create MANUAL_RULE_PRE_AGGREGATION job

- Add a new job `MANUAL_RULE_PRE_AGGREGATION` for rebuilding the
  aggregation variables of all the active rule instances
- When `MANUAL_PRE_AGGREGATION` feature flag is enabled, we don't
  change rule instance to PENDING status.

This is for PNB backfill. We'll activate all the rule instances first,
and then trigger the MANUAL_RULE_PRE_AGGREGATION job to rebuild the
aggregation variables. After the rebuild is ready, we can start to
verify the last 2-week transactions.