Fix transaction double counting issue (#5370)
We only set V8_LOGIC_USER_TIME_AGGREGATION_TX_MARKER for a transaction
when doing delta update. When we rebuild an aggregation variable, we don't
set the tx marker. So when a transaction is created before the rebuild and
a transaction event is created after the rebuild, double-counting will occur.

Change to determine if a transaction is applied before by checking if the
tx marker is set OR lastTransactionTimestamp (set by rebuild) is >= the
current transaction.