Handle RebuildSyncRetryError (#5435)
When RULES_ENGINE_V8_SYNC_REBUILD is enabled, we rebuild aggregation variable
synchronously if aggregation variable is not ready. If we still get RebuildSyncRetryError
after the rebuild, we just log error and return null. It's possible that we saw this
behaviour because the ready indicator is not immediately visible after write, thus
use ConsistenRead for it.