Make BACKFILL_ASYNC_RULE_RUNS job deterministic (#5620)
Previously, when rerunning async rules for the past transactions, we updated the
versions of the aggregation variables first. But it has 2 problems
- If the backfill runs first, new transactions will use the wrong data because
  missing aggregation data outside of the backfill time window
- If new transactions of a user come before backfill, backfill result will be wrong
  as it's using the latest aggregation data which doesn't include the backfill window

Change to add backfill job ID to the suffix of the aggregation variable version to make
it isolate from the live ones. This requires the time window to be continuous, thus
changing `affectedExecutionRanges` to be `affectedExecutionRange`.